wordstocksite/requirements.txt
amqp==5.1.1
asgiref==3.5.2
autopep8==2.0.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
dnspython==2.2.1
eventlet==0.33.2
flower==1.2.0
greenlet==2.0.1
humanize==4.4.0
idna==3.4
kombu==5.2.4
prometheus-client==0.15.0
prompt-toolkit==3.0.32
pycodestyle==2.9.1
python-decouple==3.6
pytz==2022.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
tzdata==2022.5
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
